requirements.txt
openai-whisper
sounddevice
pynput
pyperclip
toml 
Flask 
requests 
scipy 
simpleaudio 